HorseIdentityBackend/requirements.txt
fastapi
uvicorn
torch
Pillow
numpy
git+https://github.com/openai/CLIP.git
